backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
scikit-learn==1.6.1
xgboost==2.1.4
PyJWT==2.10.1
numpy==2.2.3
